Unconsolidate TransactionPool from Consensus service
