+++ projects/architecture/interfaces/consensus-service.proto
@@ -2,3 +2,4 @@ syntax = "proto3";
 
 import "./consensus.proto";
 import "./subscription-manager.proto";
+import "./transaction-pool.proto";
