+++ projects/services/consensus-service-typescript/src/index.ts
@@ -3,6 +3,7 @@ import {  } from "orbs-core-library";
 
 import ConsensusService from "./consensus-service";
 import SubscriptionManagerService from "./subscription-manager-service";
+import TransactionPoolService from "./transaction-pool-service";
 
 ErrorHandler.setup();
 
@@ -10,5 +11,5 @@ const nodeTopology = topology();
 
 grpc.consensusServiceServer({
   endpoint: nodeTopology.endpoint,
-  services: [new ConsensusService(), new SubscriptionManagerService()]
+  services: [new ConsensusService(), new SubscriptionManagerService(), new TransactionPoolService()]
 });
