+++ projects/services/consensus-service-typescript/src/transaction-pool-service.ts
@@ -1,0 +1,61 @@
+import * as _ from "lodash";
+import bind from "bind-decorator";
+
+import { logger, config, topology, topologyPeers, grpc, types } from "orbs-core-library";
+
+import { TransactionPool } from "orbs-core-library";
+
+const nodeTopology = topology();
+
+export default class TransactionPoolService {
+  private transactionPool: TransactionPool;
+
+  @bind
+  public async getHeartbeat(rpc: types.GetHeartbeatContext) {
+    logger.debug(`${nodeTopology.name}: service '${rpc.req.requesterName}(v${rpc.req.requesterVersion})' asked for heartbeat`);
+
+    rpc.res = { responderName: nodeTopology.name, responderVersion: nodeTopology.version };
+  }
+
+  @bind
+  public async addNewPendingTransaction(rpc: types.AddNewPendingTransactionContext) {
+    logger.info(`${nodeTopology.name}: addNewPendingTransaction ${JSON.stringify(rpc.req.transaction)}`);
+
+    this.transactionPool.addNewPendingTransaction(rpc.req.transaction);
+
+    rpc.res = {};
+  }
+
+  @bind
+  public async addExistingPendingTransaction(rpc: types.AddExistingPendingTransactionContext) {
+    logger.info(`${nodeTopology.name}: addExistingPendingTransaction ${JSON.stringify(rpc.req.transaction)}`);
+
+    this.transactionPool.addExistingPendingTransaction(rpc.req.transaction);
+
+    rpc.res = {};
+  }
+
+  async askForHeartbeat(peer: types.HeardbeatClient) {
+    const res = await peer.getHeartbeat({ requesterName: nodeTopology.name, requesterVersion: nodeTopology.version });
+    logger.debug(`${nodeTopology.name}: received heartbeat from '${res.responderName}(v${res.responderVersion})'`);
+  }
+
+  askForHeartbeats() {
+  }
+
+  async initTransactionPool(): Promise<void> {
+    this.transactionPool = new TransactionPool();
+  }
+
+  async main() {
+    logger.info(`${nodeTopology.name}: service started`);
+
+    await this.initTransactionPool();
+
+    setInterval(() => this.askForHeartbeats(), 5000);
+  }
+
+  constructor() {
+    setTimeout(() => this.main(), 0);
+  }
+}
